requirements/tests.txt
bandit==1.5.1
flake8-blind-except==0.1.1
flake8-coding==1.3.1
flake8-commas==2.0.0
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-isort==2.5
flake8-pep3101==1.2.1
flake8-quotes==1.0.0
flake8-string-format==0.2.3
flake8==3.6.0
ipdb==0.11
isort==4.3.4
lizard==1.15.6
pep257==0.7.0
pep8-naming==0.7.0
pep8==1.7.1
pip-check==2.3.3
pycodestyle==2.4.0
pyflakes==2.0.0
pytest-cache==1.0
pytest-cov==2.5.1
pytest-django==3.4.3
pytest==3.10.1
radon==2.4.0
yamllint==1.12.1
